Type fixes
