+++ packages/theme-check-common/src/types/schemas/index.ts
@@ -1,3 +1,4 @@
 export { ThemeBlock } from './theme-block';
 export { Section } from './section';
 export { Setting } from './setting';
+export { Preset } from './preset';
