+++ packages/theme-check-common/src/types/schemas/theme-block.ts
@@ -1,4 +1,5 @@
 import { Setting } from './setting';
+import { Preset } from './preset';
 
 export declare namespace ThemeBlock {
   /** Typed content of {% schema %} after passing validation */
@@ -6,7 +7,7 @@ export declare namespace ThemeBlock {
     name?: string;
     settings?: Setting.Any[];
     blocks?: Block[];
-    presets?: Preset[];
+    presets?: Preset.Preset[];
     class?: string;
     tag?: string | null;
   }
@@ -30,39 +31,4 @@ export declare namespace ThemeBlock {
     | AppBlock
     | SpecificThemeBlock
   );
-
-  // The preset type that uses the above
-  export interface Preset {
-    name: string;
-    settings?: Setting.Values;
-    blocks?: BlockPresets;
-  }
-
-  // Union type that represents either format
-  export type BlockPresets = BlockPresetArrayElement[] | BlockPresetHash;
-
-  // Hash format for blocks
-  export type BlockPresetHash = {
-    [id: string]: BlockPresetBase & {
-      blocks?: BlockPresets;
-    };
-  };
-
-  // Array element format for blocks
-  export type BlockPresetArrayElement = (BlockPresetBase | BlockPresetStatic) & {
-    blocks?: BlockPresets; // Recursive
-  };
-
-  /** Base type for presets */
-  export type BlockPresetBase = {
-    /** Refers to a block type. blocks/{type}.liquid */
-    type: string;
-    settings?: Setting.Values;
-  };
-
-  /** Static blocks */
-  export type BlockPresetStatic = BlockPresetBase & {
-    static: true;
-    id: string;
-  };
 }
